requirements.txt
alembic==1.14.1
Authlib==1.4.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-oidc==2.2.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
pdfplumber
instructor
openai
python-docx